Added user login, registration, view, edit, delete posts by created user

+++ www/app/Http/Controllers/PostsController.php
@@ -3,10 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -41,13 +52,32 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle file upload
+        if($request->hasFile('cover_image')){
+            // Get file with the extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just file name
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $fileExtension = $request->file('cover_image')->getClientOriginalExtension();
+            // File name to store
+            $fileNameToStore = $fileName.'_'.time().'.'.$fileExtension;
+            // Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         // Create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
         
         return redirect('/posts')->with('success', 'Post Created');
@@ -74,6 +104,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        
+        // Check for correct user
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error', 'Unathorized user to access this page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -91,10 +127,27 @@ class PostsController extends Controller
             'body' => 'required'
         ]);
 
+         // Handle file upload
+        if($request->hasFile('cover_image')){
+            // Get file with the extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just file name
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $fileExtension = $request->file('cover_image')->getClientOriginalExtension();
+            // File name to store
+            $fileNameToStore = $fileName.'_'.time().'.'.$fileExtension;
+            // Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
         // Create post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
         
         return redirect('/posts')->with('success', 'Post Updated');
@@ -109,6 +162,18 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error', 'Unathorized user to access this page');
+        }
+
+        // Check if not default image
+        if($post->cover_image != 'noimage.jpg'){
+            // Delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Removed');
